app/requirements.txt
Flask~=2.0.3
Flask-Cors~=3.0.10
python-dotenv~=0.19.2
sshtunnel~=0.4.0
mysql-connector-python~=8.0.28